+++ middleware/not-found-handler.js
@@ -1,12 +1,9 @@
-const { NotFoundError } = require('../../src')
+const { NotFoundError } = require('../http-errors')
 
 /**
  * If this middleware is reached, it means that the route does not exist
  * and a NotFoundError should be thrown
- * 
- * @param {import('express').Request} req
- * @param {import('express').Response} res
- * @param {import('express').NextFunction} next
+ * @type {import('express').RequestHandler}
  * @returns {void}
  * @throws {NotFoundError}
  *
